Added Move Counter
I'm ashamed because it took so long, but I'm proud I
learned something. When assigning a variable and doing
the whole document.getElement thing, if you add .textContent
to the end if that, it will always be the original text content,
it doesn't allow to change dynamically.

NEXT TO-DO: Star Rating thing